backend/requirements.txt
Flask
flask-cors
python-socketio
eventlet
gunicorn
supabase
python-dotenv
textblob==0.17.1
redis
requests
pydantic
Flask-Limiter==3.5.0
better-profanity==0.7.0
langdetect==1.0.9
nltk==3.8.1
flask-socketio==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0
eventlet>=0.35.2
setuptools
apscheduler==3.10.4
